Initial POC for pattern language working
